requirements.txt
openai
python-dotenv
pytesseract
Pillow